Removed VBattAdcThreshold in favour of a direct voltage threshold which is easier to understand.
Added iSCritical to publishStatus so that a delay can be added to the publishing to ensure the message has gone (pending Photon firmware update that includes confirmation).

Added flag to publishStatus to indicate if a published message may be waiting to allow the deep sleep to be delayed so the messages can be published by the photon.

Stopped roll monitoring and publishing when the battery level is low to slow down draining the batteries which can be bad for them.